frontend/requirements.txt
streamlit
requests
streamlit-extras
python-dotenv
sseclient-py
